20240503 update final

+++ src/main/java/com/prj1/controller/MemberController.java
@@ -1,9 +1,9 @@
 package com.prj1.controller;
 
 import com.prj1.domain.Member;
-import com.prj1.service.CustomUserDetailsService;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -45,10 +45,12 @@ public class MemberController {
 	}
 
 	@PostMapping("remove")
-	public String remove(Integer id) {
-		service.remove(id);
+	public String remove(Integer id, Authentication authentication) {
+		if (service.hasAccess(id, authentication)) {
+			service.remove(id);
+		}
 
-		return "redirect:/member/signup";
+		return "redirect:logout";
 	}
 
 	@GetMapping("modify")
@@ -59,9 +61,10 @@ public class MemberController {
 	}
 
 	@PostMapping("modify")
-	public String modify(Member member, RedirectAttributes rttr) {
-		service.modify(member);
-
+	public String modify(Member member, Authentication authentication, RedirectAttributes rttr) {
+		if(service.hasAccess(member.getId(), authentication)) {
+			service.modify(member);
+		}
 		rttr.addAttribute("id", member.getId());
 		return "redirect:/member";
 	}
@@ -77,11 +80,5 @@ public class MemberController {
 	@GetMapping("login")
 	public String loginForm() {
 		return "member/login";
-	}
-
-	@GetMapping("logout")
-	public String login() {
-
-		return "/";
 	}
 }
